refactor: remove utility_base
