+++ src/libcommon/utility/utility.h
@@ -20,7 +20,6 @@
 
 #include "libcommon/libcommon.h"
 #include "types.h"
-#include "utility_base.h"
 
 #include <string>
 #include <thread>
@@ -143,18 +142,15 @@ struct COMMON_EXPORT CommonUtility {
         static void writeSignalFile(AppType appType, SignalType signalType) noexcept;
         static void clearSignalFile(AppType appType, SignalCategory signalCategory, SignalType &signalType) noexcept;
 
-        static bool isLikeFileNotFoundError(const std::error_code &ec) noexcept {
-            return utility_base::isLikeFileNotFoundError(ec);
-        };
-
 #ifdef _WIN32
         // Converts a std::wstring to std::string assuming that it contains only mono byte chars
         static std::string toUnsafeStr(const SyncName &name);
 
-        static std::wstring getErrorMessage(DWORD errorMessageId) { return utility_base::getErrorMessage(errorMessageId); }
-        static std::wstring getLastErrorMessage() { return utility_base::getLastErrorMessage(); };
-        static bool isLikeFileNotFoundError(DWORD dwError) noexcept { return utility_base::isLikeFileNotFoundError(dwError); };
+        static std::wstring getErrorMessage(DWORD errorMessageId);
+        static std::wstring getLastErrorMessage();
+        static bool isLikeFileNotFoundError(DWORD dwError) noexcept;
 #endif
+        static bool isLikeFileNotFoundError(const std::error_code &ec) noexcept;
 
         static QString truncateLongLogMessage(const QString &message);
