+++ src/libcommon/utility/utility_win.cpp
@@ -110,4 +110,38 @@ bool CommonUtility::normalizedSyncName(const SyncName &name, SyncName &normalize
     return true;
 }
 
+std::string CommonUtility::toUnsafeStr(const SyncName &name) {
+    std::string unsafeName(name.begin(), name.end());
+    return unsafeName;
+}
+
+std::wstring CommonUtility::getErrorMessage(DWORD errorMessageID) {
+    LPWSTR messageBuffer = nullptr;
+    const size_t size =
+            FormatMessageW(FORMAT_MESSAGE_ALLOCATE_BUFFER | FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS, nullptr,
+                           errorMessageID, NULL, (LPWSTR) &messageBuffer, 0, nullptr);
+
+    // Escape quotes
+    const auto msg = std::wstring(messageBuffer, size);
+    std::wostringstream message;
+    message << L"(" << errorMessageID << L") - " << msg;
+
+    LocalFree(messageBuffer);
+
+    return message.str();
+}
+
+std::wstring CommonUtility::getLastErrorMessage() {
+    return getErrorMessage(GetLastError());
+}
+
+bool CommonUtility::isLikeFileNotFoundError(DWORD dwError) noexcept {
+    return (dwError == ERROR_FILE_NOT_FOUND) || (dwError == ERROR_PATH_NOT_FOUND) || (dwError == ERROR_INVALID_DRIVE) ||
+           (dwError == ERROR_BAD_NETPATH);
+}
+
+bool CommonUtility::isLikeFileNotFoundError(const std::error_code &ec) noexcept {
+    return isLikeFileNotFoundError(static_cast<DWORD>(ec.value()));
+}
+
 } // namespace KDC
